+++ packages/auth/src/Api/ActivateState/interfaces.ts
@@ -1,10 +1,12 @@
 export enum ActivateStep {
   'activating' = 'activating',
   'success' = 'success',
+  'resend' = 'resend',
 }
 
 export interface ActivateState {
   loading: boolean;
   error?: any;
   step: ActivateStep;
+  resentEmail?: boolean;
 }