+++ packages/auth/src/Api/ActivateState/saga.ts
@@ -1,7 +1,7 @@
 import { call, put, takeLeading } from 'redux-saga/effects';
 import { actions } from '../reducer';
 import { PayloadAction } from '@reduxjs/toolkit';
-import { api, IActivateAccount } from '@frontegg/rest-api';
+import { api, IActivateAccount, IResendActivationEmail } from '@frontegg/rest-api';
 import { ActivateStep } from './interfaces';
 
 function* activateAccount({ payload }: PayloadAction<IActivateAccount>) {
@@ -14,6 +14,17 @@ function* activateAccount({ payload }: PayloadAction<IActivateAccount>) {
   }
 }
 
+function* resendActivationEmailFunction({ payload }: PayloadAction<IResendActivationEmail>) {
+  yield put(actions.setActivateState({ loading: true }));
+  try {
+    yield call(api.auth.resendActivationEmail, { email: payload.email });
+    yield put(actions.setActivateState({ loading: false, error: undefined, resentEmail: true }));
+  } catch (e) {
+    yield put(actions.setActivateState({ loading: false, error: e.message }));
+  }
+}
+
 export function* activateSagas() {
   yield takeLeading(actions.activateAccount, activateAccount);
+  yield takeLeading(actions.resendActivationEmail, resendActivationEmailFunction);
 }